Configure QoS for MQTT Consumer
This commit configures the QoS policy for the MQTT consumer. Previously,
the QoS level was set to 0 (fire and forget). With the MQTT consumer
having a fast network connection to the MQTT broker we can set the
QoS level to 2 in order to avoid discarding values on the consumer side.