Allow upscaling the persistence service
Previously, the persistence service was directly consuming messages
from the multicast queue which meant that only one persistence service
could be active in order to avoid slowing down the performance due to
many duplicate inserts.

This commit introduces a forwarding of the messages received on the
multicast queue to an anycast queue. Therefore, many instances of the
persistence service can now be started.